AK+LibC: Remove SERENITY_LIBC_BUILD guard around `<initializer_list>`
This was required before commit 38948bb104dfcfaf6bb044e64fa311daf5c39bd7
when we were building LibC before libstdc++ headers were available in
the sysroot. However as noted in that commit, we never actually needed
to be building LibC before libstdc++, so we can go ahead and remove this
ancient hack.